# specs/001-package-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Elsa Platform API
  version: 0.1.0
  description: Public and admin REST contracts for package, feature, compatibility, sync, validation, and approval workflows.
servers:
  - url: http://localhost:5000
security: []
...
paths:
...
  /api/admin/sources:
    get:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: List package sources
      responses:
        '200':
          description: Package sources.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminPackageSource'
    post:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: Create package source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPackageSourceRequest'
      responses:
        '201':
          description: Created source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackageSource'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/sources/{id}:
    put:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: Update package source
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPackageSourceRequest'
      responses:
        '200':
          description: Updated source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackageSource'
    delete:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: Delete package source
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Source deleted.
...
  /api/admin/sync/sources/{sourceId}:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for one source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    SyncAccepted:
      description: Sync accepted.
      content:
        application/json:
          schema:
            type: object
            properties:
              syncRunId:
                type: string
              status:
                type: string
                enum: [Running]
...
  schemas:
...
    AdminPackageSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [NuGetFeed]
        url:
          type: string
        enabled:
          type: boolean
        includePatterns:
          type: array
          items:
            type: string
        excludePatterns:
          type: array
          items:
            type: string
        approvalPolicy:
          type: string
          enum: [AutoApprove, Manual]
        versionDiscoveryPolicy:
          $ref: '#/components/schemas/PackageSourceVersionDiscoveryPolicy'
        lastSyncedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpsertPackageSourceRequest:
      type: object
      required: [name, type, url, enabled, includePatterns, approvalPolicy]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [NuGetFeed]
        url:
          type: string
        enabled:
          type: boolean
        includePatterns:
          type: array
          items:
            type: string
        excludePatterns:
          type: array
          items:
            type: string
        approvalPolicy:
          type: string
          enum: [AutoApprove, Manual]
        versionDiscoveryPolicy:
          type: string
          enum: [AllVersions, LatestStable, LatestIncludingPrerelease, LatestPreview]
          default: AllVersions
    PackageSourceVersionDiscoveryPolicy:
      type: string
      enum: [AllVersions, LatestStable, LatestIncludingPrerelease, LatestPreview]
    SyncRequest:
      type: object
      properties:
        forceReindex:
          type: boolean
          default: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string